requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
